+++ options.go
@@ -1,62 +1,93 @@
 package ginzerologger
 
-type StatusLevel int
+import "github.com/rs/zerolog"
 
-const (
-	Above200 StatusLevel = 200
-	Above300 StatusLevel = 300
-	Above400 StatusLevel = 400
-	Above500 StatusLevel = 500
-)
+type HTTPStatus int
 
-var LogBodyErrorStatusLevel = struct {
-	Above200 StatusLevel
-	Above300 StatusLevel
-	Above400 StatusLevel
-	Above500 StatusLevel
+var HTTPStatusCodes = struct {
+	EqualToOrGreaterThan200 HTTPStatus
+	EqualToOrGreaterThan300 HTTPStatus
+	EqualToOrGreaterThan400 HTTPStatus
+	EqualToOrGreaterThan500 HTTPStatus
 }{
-	Above200: Above200,
-	Above300: Above300,
-	Above400: Above400,
-	Above500: Above500,
+	EqualToOrGreaterThan200: 2,
+	EqualToOrGreaterThan300: 3,
+	EqualToOrGreaterThan400: 4,
+	EqualToOrGreaterThan500: 5,
 }
 
-type LoggingOption struct {
+type logLevel interface {
+	*zerolog.Event | string
+}
+
+type loggingOption struct {
 	Key   string
 	Value any
 }
 
-func newLoggingOption(key string, value any) LoggingOption {
-	return LoggingOption{
+func newLoggingOption(key string, value any) *loggingOption {
+	return &loggingOption{
 		Key:   key,
 		Value: value,
 	}
 }
 
-func LogRequestBodyOption(lvl StatusLevel) LoggingOption {
-	return newLoggingOption("logRequestBody", lvl)
+func IncludeRequestBody(sts HTTPStatus) *loggingOption {
+	return newLoggingOption("includeRequestBody", sts)
 }
 
-func LogPathExclusion(path string) LoggingOption {
-	return newLoggingOption("exclude", path)
+func LogLevelEqualToOrGreaterThan200[T logLevel](val T) *loggingOption {
+	return &loggingOption{
+		Key:   "default200",
+		Value: val,
+	}
+}
+
+func LogLevelEqualToOrGreaterThan300[T logLevel](val T) *loggingOption {
+	return &loggingOption{
+		Key:   "default300",
+		Value: val,
+	}
+}
+
+func LogLevelEqualToOrGreaterThan400[T logLevel](val T) *loggingOption {
+	return &loggingOption{
+		Key:   "default400",
+		Value: val,
+	}
+}
+
+func LogLevelEqualToOrGreaterThan500[T logLevel](val T) *loggingOption {
+	return &loggingOption{
+		Key:   "default500",
+		Value: val,
+	}
+}
+
+func PathExclusion(paths ...string) *loggingOption {
+	if len(paths) > 0 {
+		return newLoggingOption("excludes", paths)
+	}
+
+	return nil
 }
 
 type optionsSearch struct {
-	opts []LoggingOption
+	opts []*loggingOption
 }
 
-func newOptionsSearch(opts ...LoggingOption) *optionsSearch {
+func newOptionsSearch(opts ...*loggingOption) *optionsSearch {
 	return &optionsSearch{
 		opts: opts,
 	}
 }
 
-func (o *optionsSearch) Find(key string) (LoggingOption, bool) {
+func (o *optionsSearch) Find(key string) (*loggingOption, bool) {
 	for _, opt := range o.opts {
 		if opt.Key == key {
 			return opt, true
 		}
 	}
 
-	return LoggingOption{}, false
+	return nil, false
 }