+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2023-06-22
 
-- Adding logging options to `GinZeroLogger()` function to further tailor how request logging is handled
+- Added logging options to `GinZeroLogger()` function to further tailor how request logging is handled
+- Added documentation to the README.md
+- Added `example/main.go` to demonstrate usage of the module
 
 ## [1.0.0] - 2023-06-21